+++ src/main.rs
@@ -164,9 +164,10 @@ impl libcnb::Buildpack for NodeJsBuildpack {
         ))
         .inspect(package_manager::log_requested_package_manager)
         .and_then(|requested_package_manager| {
-            package_manager::resolve_package_manager(&context, &mut env, &requested_package_manager)
+            package_manager::resolve_package_manager(&context, &env, &requested_package_manager)
         })
         .inspect(package_manager::log_resolved_package_manager)
+        .and_then(package_manager::check_package_manager_support_status)
         .and_then(|resolved_package_manager| {
             package_manager::install_package_manager(&context, &mut env, &resolved_package_manager)
         })?;